refactor: moves the xAPI transforms to their own module
